requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
diff-match-patch==20241021
Django==5.1.6
django-browser-reload==1.18.0
django-countries==7.6.1
django-import-export==4.3.7
django-tailwind==3.8.0
idna==3.10
isort==6.0.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.1.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
